+++ app/akademik/page.jsx
@@ -1,4 +1,8 @@
-export default function Akademik() {
+import Navbar from "@/components/navbar/Navbar";
+import Footer from "@/components/footer/Footer";
+import Akademik from "@/features/akademik/Akademi";
+
+export default function PageAkademik() {
     return (
       <div>
         <Navbar />
