+++ app/berita/page.jsx
@@ -1,4 +1,8 @@
-export default function Berita() {
+import Navbar from "@/components/navbar/Navbar";
+import Footer from "@/components/footer/Footer";
+import Berita from "@/features/berita/Berita";
+
+export default function PageBerita() {
     return (
       <div>
         <Navbar />
