+++ app/kontak/page.jsx
@@ -1,4 +1,8 @@
-export default function Kontak() {
+import Navbar from "@/components/navbar/Navbar";
+import Footer from "@/components/footer/Footer";
+import Kontak from "@/features/kontak/Kontak";
+
+export default function PageKontak() {
     return (
       <div>
         <Navbar />
