+++ app/prestasi/page.jsx
@@ -1,4 +1,8 @@
-export default function Prestasi() {
+import Navbar from "@/components/navbar/Navbar";
+import Footer from "@/components/footer/Footer";
+import Prestasi from "@/features/prestasi/Prestasi";
+
+export default function PagePrestasi() {
     return (
       <div>
         <Navbar />
