+++ app/profile/page.jsx
@@ -1,4 +1,8 @@
-export default function Profile() {
+import Navbar from "@/components/navbar/Navbar";
+import Footer from "@/components/footer/Footer";
+import Profile from "@/features/profile/Profile";
+
+export default function PageProfile() {
     return (
       <div>
         <Navbar />
