+++ features/akademik/Akademi.jsx
@@ -1,0 +1,10 @@
+import Image from "next/image";
+import Link from "next/link";
+
+export default function Akademik() {
+  return (
+    <div className="w-full h-[500px]">
+        <p>ini akademik</p>
+    </div>
+  );
+}
