+++ features/berita/Berita.jsx
@@ -1,0 +1,10 @@
+import Image from "next/image";
+import Link from "next/link";
+
+export default function Berita() {
+  return (
+    <div className="w-full h-[500px]">
+        <p>ini Berita</p>
+    </div>
+  );
+}
