+++ features/kontak/Kontak.jsx
@@ -1,0 +1,10 @@
+import Image from "next/image";
+import Link from "next/link";
+
+export default function Kontak() {
+  return (
+    <div className="w-full h-[500px]">
+        <p>ini Kontak</p>
+    </div>
+  );
+}
