+++ features/profile/Profile.jsx
@@ -1,0 +1,10 @@
+import Image from "next/image";
+import Link from "next/link";
+
+export default function Profile() {
+  return (
+    <div className="w-full h-[500px]">
+        <p>ini profile</p>
+    </div>
+  );
+}
